Changed the type

+++ include/lion/app.h
@@ -91,7 +91,7 @@ typedef struct lion_app_config {
   double sim_step_seconds;
   double sim_epsabs;
   double sim_epsrel;
-  double sim_min_max_iter;
+  uint64_t sim_min_max_iter;
 
   /* Logging configuration */
